Fix 4

+++ include/common.h
@@ -3,8 +3,8 @@
 #define INCLUDE_COMMON_H_
 
 struct Pair {
-  unsigned int x;
-  unsigned int y;
+  long unsigned int x;
+  long unsigned int y;
 };
 
 #endif  // INCLUDE_COMMON_H_
